// tests-unit/client.gen.ts
export interface Message {
  id: bigint
  userId: bigint
  timestamp: bigint
  content: string
}

export interface BigIntArrayTest {
  ids: Array<bigint>
  timestamps: Array<bigint>
  messages: Array<Message>
  optionalIds?: Array<bigint>
}